EmpiricalCDFs

EmpiricalCDFs

Empirical cumulative distribution functions

The source repository is https://github.com/jlapeyre/EmpiricalCDFs.jl.

Provides empirical cumulative distribution functions (CDFs) (or "empirical distribution functions" as they are know to probabalists).

EmpiricalCDFs implements empirical CDFs; building, evaluating, random sampling, evaluating the inverse, etc. It is useful especially for examining the tail of the CDF obtained from streaming a large number of data, more than can be stored in memory. For this purpose, you specify a lower cutoff; data points below this value will be silently rejected, but the resulting CDF will still be properly normalized. This ability to process and filter data online is absent in StatsBase.ecdf.

 cdf = EmpiricalCDF()
 push!(cdf,x)
 ...
 print(io,cdf)

# reject points `x < xmin` to use less memory
 cdf = EmpiricalCDF(xmin)

Warning about sorting

Before using the cdf, you must call sort!(cdf). For efficiency data is not sorted as it is inserted. The exception is print, which does sort the cdf before printing.

Contents

EmpiricalCDFs

Contents

Index

Empirical CDF types

Functions

Text file output

Binary IO

Comparison with ecdf

Index

EmpiricalCDFs.AbstractEmpiricalCDF

EmpiricalCDFs.CDFfile

EmpiricalCDFs.EmpiricalCDF

EmpiricalCDFs.EmpiricalCDFHi

Base.Random.rand

Base.append!

Base.print

Base.push!

Base.sort!

EmpiricalCDFs.counts

EmpiricalCDFs.data

EmpiricalCDFs.finv

EmpiricalCDFs.getcdf

EmpiricalCDFs.header

EmpiricalCDFs.linprint

EmpiricalCDFs.logprint

EmpiricalCDFs.readcdf

EmpiricalCDFs.readcdfinfo

EmpiricalCDFs.save

EmpiricalCDFs.version

Empirical CDF types

EmpiricalCDFs.AbstractEmpiricalCDF — Type.

AbstractEmpiricalCDF

Concrete types are EmpiricalCDF and EmpiricalCDFHi.

source

EmpiricalCDFs.EmpiricalCDF — Type.

EmpiricalCDF()

Return an empirical CDF. After inserting elements with push! or append!, and before being accessed using any of the functions below, the CDF must be sorted with sort!.

EmpiricalCDF is a callable object:

(cdf::EmpiricalCDF)(x::Real)

returns the value of the approximate cummulative distribution function cdf at the point x.

source

EmpiricalCDF(lowreject::Real)

If lowereject is finite return EmpiricalCDFHi(lowreject). Otherwise return EmpiricalCDF().

source

EmpiricalCDFs.EmpiricalCDFHi — Type.

EmpiricalCDFHi{T <: Real} <: AbstractEmpiricalCDF

Empirical CDF with lower cutoff. That is, keep only the tail.

source

Functions

Base.push! — Function.

push!(cdf::EmpiricalCDF,x::Real)

add the sample x to cdf.

source

Base.append! — Function.

append!(cdf::EmpiricalCDF, a::AbstractArray)

add samples in a to cdf.

source

Base.sort! — Function.

sort!(cdf::AbstractEmpiricalCDF)

Sort the data collected in cdf. You must call sort! before using cdf.

source

EmpiricalCDFs.data — Function.

data(cdf::AbstractEmpiricalCDF)

return the array holding samples for cdf.

source

EmpiricalCDFs.counts — Function.

counts(cdf::AbstractEmpiricalCDF)

return the number of counts added to cdf. This includes counts that may have been discarded because they are below of the cutoff.

source

Base.Random.rand — Function.

rand(cdf::EmpiricalCDF)

Pick a random sample from the distribution represented by cdf.

source

EmpiricalCDFs.finv — Function.

finv(cdf::AbstractEmpiricalCDF) --> Function

Return the functional inverse of cdf. cdf is a callable object. finv(cdf) returns a function that captures cdf in a closure.

source

Methods are defined on AbstractEmpiricalCDF for the following functions: length, minimum, maximum, extrema, mean, median, std, quantile.

Text file output

Base.print — Function.

print(io::IO, cdf::AbstractEmpiricalCDF)

Call logprint(io,cdf)

source

EmpiricalCDFs.linprint — Function.

linprint(io::IO ,cdf::AbstractEmpiricalCDF, n=2000) print (not more than) n linearly spaced points after sorting the data.

source

linprint(fn::String, cdf::AbstractEmpiricalCDF, n=2000)

print cdf to file fn. Print no more than n linearly spaced points.

source

EmpiricalCDFs.logprint — Function.

logprint(io::IO, cdf::EmpiricalCDF, n=2000) print (not more than) n log spaced points after sorting the data.

source

Binary IO

I found available serialization choices to be too slow. So, very simple, very fast, binary storage and retrieval is provided. By now, or in the future, there will certainly be packages that provide a sufficient or better replacement.

The type CDFfile supports reading and writing AbstractEmpiricalCDF objects in binary format. Most functions that operate on AbstractEmpiricalCDF also work with CDFfile, with the call being passed to the cdf field.

EmpiricalCDFs.CDFfile — Type.

CDFfile(cdf::AbstractEmpiricalCDF, header="")

struct CDFfile{T <: AbstractEmpiricalCDF}
    cdf::T
    header::String
end

Binary data file for AbstractEmpiricalCDF

The file format is

Identifying string

n::Int64 number of bytes in the header string

s::String The header string

t::Int64 Type of AbstractEmpiricalCDF, 1 or 2. 1 for EmpiricalCDF, 2 for EmpiricalCDFHi.

lowreject::Float64 the lower cutoff, only for EmpiricalCDFHi.

npts::Int64 number of data points in the CDF

npts data points of type Float64

source

EmpiricalCDFs.save — Function.

save(fn::String, cdf::AbstractEmpiricalCDF, header::String="")

write cdf to file fn in a fast binary format.

source

EmpiricalCDFs.readcdf — Function.

readcdf(fn::String)

Read an empirical CDF from file fn. Return an object cdff of type CDFfile. The header is in field header. The cdf is in in field cdf.

source

EmpiricalCDFs.readcdfinfo — Function.

readcdfinfo(fn::String)

return an object containing information about the cdf saved in the binary file fn. The data itself is not read.

source

EmpiricalCDFs.header — Function.

header::String = header(cdff::CDFfile)

Return the header from cdff.

source

EmpiricalCDFs.getcdf — Function.

cdf::AbstractEmpiricalCDF = getcdf(cdff::CDFfile)

Return the CDF from cdff.

source

EmpiricalCDFs.version — Function.

version(cdff::CDFfile)

Return the version number of the file format.

source

Comparison with ecdf

This package differs from the ecdf function from StatsBase.jl.

ecdf takes a sorted vector as input and returns a function that looks up the value of the CDF. An instance of EmpiricalCDF, cdf, both stores data, eg via push!(cdf,x), and looks

up the value of the CDF via cdf(x).

When computing the CDF at an array of values, ecdf, sorts the input and uses an algorithm that scans the data. Instead, EmpiricalCDFs does a binary search for each element of an input vector. Tests showed that this is typically not slower. If the CDF stores a large number of points relative to the size of the input vector, the second method, the one used by EmpiricalCDFs is faster.